+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-26 22:34:43
-// Commit: feat(network): add bi-directional event bus handler for network
+// Updated at: 2026-05-27 10:45:51
+// Commit: chore(network): update dependencies and cleanup unused imports in network
 
 class networkComponentState {
-  final String commitRef = '1479';
-  final String lastMessage = 'feat(network): add bi-directional event bus handler for network';
-  final DateTime timestamp = DateTime.parse('2026-05-26T22:34:43+05:00');
+  final String commitRef = '1484';
+  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
+  final DateTime timestamp = DateTime.parse('2026-05-27T10:45:51+05:00');
 }